Сделал разное отображение карточек в блогах на разных разрешениях. Почти закончил верстку. + добавить TS
